requirement.txt
dagster 
duckdb 
dagster-dbt 
dbt-duckdb
plotly 
pandas
psycopg2-binary 